Split editor-header into visual and logic
